fix(backend) Update Contact
Update Contact to fix the error that the contact was not being saved in the database

+++ backend/src/main/java/com/CRM_Entertainment/backend/Models/DAO/ContactDaoImp.java
@@ -11,52 +11,40 @@ import jakarta.persistence.PersistenceContext;
 import jakarta.transaction.Transactional;
 
 @Repository
-
 public class ContactDaoImp implements IContactDao {
 
     @PersistenceContext
-    
     private EntityManager em;
 
     @SuppressWarnings("unchecked")
     @Transactional
     @Override
-    public List<Contact> findAll() 
-    { 
-      return em.createQuery("from Contact").getResultList();
+    public List<Contact> findAll() {
+        return em.createQuery("from Contact").getResultList();
     }
 
     @Override
     @Transactional
-    public void save(Contact contact)
-    {
-        if(contact.getId() != null && contact.getId() > 0)
-        {
-          em.merge(contact);
+    public void save(Contact contact) {
+        if (contact.getId() != null && contact.getId() > 0) {
+            em.merge(contact);
+        } else {
+            em.persist(contact);
         }
-        else
-        {
-          em.persist(contact);
+    }
+
+    @Override
+    @Transactional
+    public Contact findOne(Long id) {
+        return em.find(Contact.class, id);
+    }
+
+    @Override
+    @Transactional
+    public void delete(Long id) {
+        Contact contact = findOne(id);
+        if (contact != null) {
+            em.remove(contact);
         }
-        
     }
-
-
-    @Override
-    @Transactional
-    public Contact findOne(Long Id) {
-
-      return em.find(Contact.class, Id);
-    }
-
-    @Override
-    @Transactional
-    public void delete(Long Id) {
-      
-      Contact contact = findOne(Id);
-
-      em.remove(contact);
-
-    }
-
-  }
+}
